refactor (User): User class move to user package
